Far 64 admin pantallas

+++ src/main/java/farmappceuticos/farmappcia/controller/EventController.java
@@ -1,7 +1,14 @@
 package farmappceuticos.farmappcia.controller;
+import farmappceuticos.farmappcia.model.Agenda;
 import farmappceuticos.farmappcia.model.Event;
+import farmappceuticos.farmappcia.model.User;
 import farmappceuticos.farmappcia.services.EventService;
+import farmappceuticos.farmappcia.services.UserService1;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PostAuthorize;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -9,12 +16,16 @@ import org.springframework.web.bind.annotation.*;
 import java.util.Optional;
 
 @Controller
-@RequestMapping("/eventos")//url
+@RequestMapping("/evento")//url
 public class EventController {
     @Autowired
     private EventService eventService;
+
+    @Autowired
+    private UserService1 userService;
     //Para acceder a los métodos
 
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @GetMapping("/")
     //Model es el objeto que utiliza Spring para pasar al html los datos de la BD
     public String showEvents(Model model){
@@ -23,37 +34,71 @@ public class EventController {
         //Devuelve el HTML
         return "event/event-list";
     }
-    @GetMapping("/new")
-    public String showNewEventForm(Model model) {
-        model.addAttribute("event", new Event());
-        return "event/event-form";
-    }
+
     @PostMapping("/save")
     public String saveEvent(@ModelAttribute("event") Event event) throws Exception {
-        eventService.save(event);
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UserDetails userDetails = null;
+        if (principal instanceof UserDetails) {
+            userDetails = (UserDetails) principal;
+        }
 
-        return "redirect:/eventos/";
+        String userName = userDetails.getUsername();
+        User user = userService.findByName(userName);
+        Agenda agenda=user.getAgendaToUser();
+        event.setDiames(event.getFechahora().getDayOfMonth());
+        event.setHora(event.getFechahora().getHour()+" "+event.getFechahora().getMinute());
+        System.out.println(event.getHora());
+       eventService.crearEvento(event,agenda);
 
-       
-
+        return "redirect:/usuario/agenda";
     }
+
     @GetMapping("/edit/{id}")
     public String showEditEventForm(@PathVariable("id") Integer id, Model model) {
-        Optional<Event> event = eventService.findById(id);
-        if (event.isPresent()){
-            model.addAttribute("event", event.get());
-            return "event/event-form";
-        }else{
-            return "error";
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UserDetails userDetails = null;
+        if (principal instanceof UserDetails) {
+            userDetails = (UserDetails) principal;
+        }
+        String userName = userDetails.getUsername();
+        User userAut = userService.findByName(userName);
+        Optional<Event> event=eventService.findById(id);
+        User user= event.get().getAgendaToEvents().getUserToAgenda();
+        if (userAut==user){
+            if (event.isPresent()){
+                model.addAttribute("event", event.get());
+                model.addAttribute("user",user);
+                return "event/event-form";
+            }else{
+                return "error";
+            }
+        }else {
+            return "error-authentication";
+        }
+    }
+
+
+    @GetMapping("/delete/{id}")
+    public String deleteEvent(@PathVariable("id") Integer id) {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UserDetails userDetails = null;
+        if (principal instanceof UserDetails) {
+            userDetails = (UserDetails) principal;
+        }
+        String userName = userDetails.getUsername();
+        User userAut = userService.findByName(userName);
+        Optional<Event> event=eventService.findById(id);
+        User user= event.get().getAgendaToEvents().getUserToAgenda();
+        if(userAut==user){
+            eventService.deleteById(id);
+            return "redirect:/usuario/agenda";
+        }else {
+            return "error-authentication";
         }
 
 
-    }
-    @GetMapping("/delete/{id}")
-    public String deleteEvent(@PathVariable("id") Integer id) {
-        eventService.deleteById(id);
 
-        return "redirect:/eventos/";
 
     }
